created mappers to convert data types

+++ app/src/main/java/com/fake/marketplace/data/source/remote/entities/FeedbackDto.kt
@@ -2,5 +2,5 @@ package com.fake.marketplace.data.source.remote.entities
 
 data class FeedbackDto(
     val count: Int,
-    val rating: Double
+    val rating: Float
 )
